# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LaraFleet API",
        "description": "Fleet management system API",
        "contact": {
            "email": "admin@larafleet.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login utente",
                "operationId": "67d6d0fdd54fba132e1b6f55ca254183",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@larafleet.test"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login effettuato"
                    },
                    "401": {
                        "description": "Credenziali non valide"
                    },
                    "422": {
                        "description": "Errore di validazione"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout utente",
                "operationId": "28d48c13fca984400ac8de99f54ee13a",
                "responses": {
                    "200": {
                        "description": "Logout effettuato"
                    },
                    "401": {
                        "description": "Non autenticato"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Utente autenticato",
                "operationId": "bc2032a379aa97707ccc61511897c120",
                "responses": {
                    "200": {
                        "description": "Dati utente corrente"
                    },
                    "401": {
                        "description": "Non autenticato"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/bookings": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Lista prenotazioni",
                "operationId": "b89b932c131e727a3720a01899acd0a2",
                "responses": {
                    "200": {
                        "description": "Lista paginata di prenotazioni"
                    },
                    "401": {
                        "description": "Non autenticato"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Crea prenotazione",
                "operationId": "166547aa028b063e31d6efe9c093aa88",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vehicle_id",
                                    "start_at",
                                    "end_at"
                                ],
                                "properties": {
                                    "vehicle_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "start_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-06-01 09:00:00"
                                    },
                                    "end_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-06-03 18:00:00"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Prenotazione creata"
                    },
                    "401": {
                        "description": "Non autenticato"
                    },
                    "422": {
                        "description": "Errore di validazione"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/bookings/{id}": {
            "delete": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Cancella prenotazione",
                "operationId": "b14ac6c3aa1a450e338f1424cc7d7168",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Prenotazione cancellata"
                    },
                    "403": {
                        "description": "Non autorizzato"
                    },
                    "404": {
                        "description": "Prenotazione non trovata"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/maintenances": {
            "get": {
                "tags": [
                    "Maintenances"
                ],
                "summary": "Lista manutenzioni",
                "operationId": "cd4ec7c16a72790431dc45429d9ebb27",
                "responses": {
                    "200": {
                        "description": "Lista paginata di manutenzioni"
                    },
                    "401": {
                        "description": "Non autenticato"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Maintenances"
                ],
                "summary": "Registra manutenzione (admin)",
                "operationId": "c7cb9353fe77fc3b44580fc6cd9e760c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vehicle_id",
                                    "description",
                                    "scheduled_at"
                                ],
                                "properties": {
                                    "vehicle_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Tagliando annuale"
                                    },
                                    "scheduled_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-15"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Manutenzione registrata"
                    },
                    "403": {
                        "description": "Non autorizzato"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/maintenances/{id}/complete": {
            "patch": {
                "tags": [
                    "Maintenances"
                ],
                "summary": "Completa manutenzione (admin)",
                "operationId": "9fe6984b6a1911597bed38f6ec58eafd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Manutenzione completata"
                    },
                    "403": {
                        "description": "Non autorizzato"
                    },
                    "404": {
                        "description": "Manutenzione non trovata"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}